+++ app/src/main/java/com/example/login/ui/viewmodels/forms/ConductorVehiculoTerceroViewModel.kt
@@ -85,7 +85,7 @@ class ConductorVehiculoTerceroViewModel (
             solicitud.conductorAfectado.datosPersona.nombreCompleto ="${campos[0].value.value} ${campos[1].value.value}"
             solicitud.conductorAfectado.datosPersona.domicilio.calle = campos[2].value.value
             solicitud.conductorAfectado.datosPersona.domicilio.numero = campos[3].value.value.toInt()
-            solicitud.conductorAfectado.datosPersona.domicilio.piso = if (campos[4].value.value.isEmpty()) null else campos[4].value.value.toInt()
+            solicitud.conductorAfectado.datosPersona.domicilio.piso = if (campos[4].value.value.isEmpty()) null else campos[4].value.value
             solicitud.conductorAfectado.datosPersona.domicilio.departamento = campos[5].value.value
             solicitud.conductorAfectado.datosPersona.domicilio.codigoPostal = campos[6].value.value.toInt()
             solicitud.conductorAfectado.datosPersona.dni= campos[7].value.value.toInt()
